requirements.txt
pandas
python-dateutil
Flask
Pillow
pytesseract
openpyxl
